refactor: Use css module on post page
