+++ src/components/post/PostPage/PostPage.module.css
@@ -1,0 +1,26 @@
+.postContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 30px 16px 80px 16px;
+  margin: 0 auto;
+}
+
+@media (min-width: 48rem) {
+  .postContainer  {
+    max-width: 768px;
+  }
+}
+
+@media (min-width: 62rem) {
+  .postContainer {
+    max-width: 992px;
+  }
+}
+
+.backButtonContainer {
+  display: flex;
+  column-gap: 6px;
+  row-gap: 4px;
+  flex-wrap: wrap;
+}
