--- src/components/post/PostPage/PostPage.styles.ts
@@ -1,25 +1,0 @@
-import styled from 'styled-components'
-import { sizes } from '@/src/theme'
-
-export const PostContainer = styled.article`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: 30px 16px 80px 16px;
-  margin: 0 auto;
-
-  @media (min-width: ${sizes.small}) {
-    max-width: 768px;
-  }
-
-  @media (min-width: ${sizes.medium}) {
-    max-width: 992px;
-  }
-`
-
-export const BackButtonContainer = styled.div`
-  display: flex;
-  column-gap: 6px;
-  row-gap: 4px;
-  flex-wrap: wrap;
-`
